requirements.txt
snowflake-connector-python
streamlit
matplotlib